packages/salesforce/pyproject.toml
[project]
name = "cogency-salesforce"
version = "0.0.1"
description = "Salesforce integration: JWT auth, Bulk 2.0, Pub/Sub gRPC CDC, writer outbox"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27",
    "simple-salesforce>=1.12",
    "grpcio>=1.66",
    "grpcio-tools>=1.66",
    "protobuf>=5.28",
    "pyjwt[crypto]>=2.9",
    "tenacity>=9.0",
    "pydantic>=2.9",
    "structlog>=24.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["salesforce"]
